refactor: create component for list of projects
