+++ src/pages/index.astro
@@ -4,7 +4,7 @@ import { getCollection } from 'astro:content';
 import Base from '@/layouts/Base.astro';
 import AboutMe from '@/partials/home/AboutMe.astro';
 import { Hero } from '@/partials/home/Hero';
-import { ProjectList } from '@/partials/home/ProjectList';
+import ProjectList from '@/partials/home/ProjectList.astro';
 import { RecentPosts } from '@/partials/home/RecentPosts';
 import { AppConfig } from '@/utils/AppConfig';
 import { sortByDate } from '@/utils/Posts';
