+++ src/pages/projects/[...page].astro
@@ -6,11 +6,9 @@ import { getCollection } from 'astro:content';
 import {
   NewerOlderPagination,
   PaginationHeader,
-  Project,
+  ProjectCatalog,
   Section,
-  Tags,
 } from '@/components';
-import { SKILL_COLOR_MAP } from '@/constants/skillTagColors';
 import Base from '@/layouts/Base.astro';
 import { AppConfig } from '@/utils/AppConfig';
 
@@ -42,28 +40,9 @@ const description = 'A page for all my notable projects';
     />
   </Section>
   <Section>
-    <div class="flex flex-col gap-6">
-      {
-        page.data.map((p) => (
-          <Project
-            name={p.data.name}
-            description={p.data.description}
-            link={p.slug}
-            img={{ src: p.data.imgSrc, alt: p.data.imgAlt }}
-            category={undefined}
-            source="https://github.com/rl16432/unlimited-wordle"
-          >
-            <Fragment slot="category">
-              {p.data.skills.map((skill) => (
-                <Tags color={SKILL_COLOR_MAP[skill]}>{skill}</Tags>
-              ))}
-            </Fragment>
-          </Project>
-        ))
-      }
-    </div>
+    <ProjectCatalog projects={page.data} />
   </Section>
   <Section>
     <NewerOlderPagination page={page} />
   </Section>
-</Base>@/constants/skillTagColors
+</Base>
